requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyJWT==2.8.0
pymongo==4.6.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
stripe==7.7.0
python-dotenv==1.0.0
google-cloud-storage==2.14.0
# Optional: For full BlueDart SOAP API integration
# zeep>=4.2.1
email-validator==2.3.0
